python: embedding generate_ocp_solver in class constructor
