+++ interfaces/acados_template/acados_template/__init__.py
@@ -39,5 +39,5 @@ from .generate_c_code_constraint import *
 from .generate_c_code_nls_cost import *
 from .AcadosOcp import *
 from .AcadosSim import *
-from .generate_ocp_solver import *
-from .generate_sim_solver import *
+from .AcadosOcpSolver import *
+from .AcadosSimSolver import *
